[Loggings] update logging logic (#54)
* [Fix] rearrange location of init eval_logger

* Ignore DeprecationWarnings in lmms_eval/__main__.py and lmms_eval/models/fuyu.py

* Update lmms_eval/__main__.py and lmms_eval/utils.py

* update

* Update llava.py with LLaVA model imports and error handling

* Add and test new datasets

* update

* Update wandb version and require report-editing:v0

* Add support for logging samples to Weights and Biases

This commit adds a new command-line argument `--wandb_log_samples` to enable logging all model outputs and documents for per-sample measurement and post-hoc analysis to Weights and Biases. The `cli_evaluate` function has been modified to handle this new argument and log the samples if the argument is set to True. The `wandb_logger` object has also been updated to include a new method `log_eval_samples` for logging the samples. This enhancement improves the functionality of the LMMS evaluation tool.

* update